Graphical representation of results
This computes the average distance between the 1st place, 2nd place, 3rd place....27th place finishers and the host city and displays it as a bar chart. This bar chart is overlayed with the average of the average for the first 7 places, the 8th-13th places, the 14th-21st places and the 22nd-27th places.